+++ src/DefaultConsumer.php
@@ -23,6 +23,8 @@ use Psr\Log\NullLogger;
  */
 final class DefaultConsumer implements Consumer
 {
+    const EXIT_ERROR = 2;
+
     /**
      * @var Driver
      */
@@ -72,7 +74,23 @@ final class DefaultConsumer implements Consumer
     {
         $this->running = true;
         while ($this->running) {
-            $this->safeOnce($queueName);
+            try {
+                $this->once($queueName);
+            } catch (Exception\MustStop $e) {
+                $this->logger->warning('Caught a must stop exception, exiting: {msg}', [
+                    'msg'   => $e->getMessage(),
+                ]);
+                $this->stop();
+                $this->exitCode = $e->getCode();
+            } catch (\Exception $e)  {
+                // likely means means something went wrong with the driver
+                $this->logger->emergency('Caught an unexpected {cls} exception, exiting: {msg}', [
+                    'cls' => get_class($e),
+                    'msg' => $e->getMessage(),
+                ]);
+                $this->stop();
+                $this->exitCode = self::EXIT_ERROR;
+            }
         }
 
         return $this->exitCode;
@@ -85,22 +103,15 @@ final class DefaultConsumer implements Consumer
     {
         $envelope = $this->driver->dequeue($queueName);
         if (!$envelope) {
-            return;
+            return null;
         }
 
+        $result = false;
         $message = $envelope->unwrap();
-        try {
-            $this->logger->debug('Executing message {msg}', ['msg' => $message->getName()]);
-            $result = $this->executor->execute($message);
-            $this->logger->debug('Executed message {msg}', ['msg' => $message->getName()]);
-        } catch (Exception\MustStop $e) {
-            // MustStop exceptions are thrown by handlers to indicate a
-            // graceful stop is required. So we don't wrapped them. Just rethrow
-            throw $e;
-        } catch (\Exception $e) {
-            $this->failed($queueName, $envelope);
-            throw new Exception\MessageFailed($e, $message);
-        }
+
+        $this->logger->debug('Executing message {msg}', ['msg' => $message->getName()]);
+        $result = $this->executeMessage($message);
+        $this->logger->debug('Executed message {msg}', ['msg' => $message->getName()]);
 
         if ($result) {
             $this->driver->ack($queueName, $envelope);
@@ -109,6 +120,8 @@ final class DefaultConsumer implements Consumer
             $this->failed($queueName, $envelope);
             $this->logger->debug('Failed message {msg}', ['msg' => $message->getName()]);
         }
+
+        return $result;
     }
 
     /**
@@ -119,37 +132,34 @@ final class DefaultConsumer implements Consumer
         $this->running = false;
     }
 
-    private function safeOnce($queueName)
-    {
-        try {
-            $this->once($queueName);
-        } catch (Exception\DriverError $e) {
-            $this->logger->critical('Caught a {cls} Driver Error, exiting: {msg}', [
-                'cls' => get_class($e),
-                'msg' => $e->getMessage(),
-            ]);
-            throw $e;
-        } catch (Exception\MustStop $e) {
-            $this->logger->warning('Caught a must stop exception, exiting: {msg}', [
-                'msg'   => $e->getMessage(),
-            ]);
-            $this->stop();
-            $this->exitCode = $e->getCode();
-        } catch (Exception\MessageFailed $e) {
-            $this->logger->critical('Unexpected {cls} exception handling {name} message: {msg}', [
-                'cls'   => get_class($e->getPrevious()),
-                'name'  => $e->getQueueMessage()->getName(),
-                'msg'   => $e->getMessage()
-            ]);
-        }
-    }
-
     private function failed($queueName, Envelope $env)
     {
         if ($this->retries->canRetry($env)) {
             $this->driver->retry($queueName, $env);
         } else {
             $this->driver->fail($queueName, $env);
+        }
+    }
+
+    private function executeMessage(Message $message)
+    {
+        try {
+            return $this->executor->execute($message);
+        } catch (Exception\MustStop $e) {
+            // MustStop exceptions are thrown by handlers to indicate a
+            // graceful stop is required. So we don't wrap them. Just rethrow
+            throw $e;
+        } catch (\Exception $e) {
+            // any other exception is simply logged. We and marked as failed
+            // below. We only log here because we can't make guarantees about
+            // the implementation of the executor and whether or not it actually
+            // throws exceptions on failure (see ForkingExecutor).
+            $this->logger->critical('Unexpected {cls} exception handling {name} message: {msg}', [
+                'cls'   => get_class($e),
+                'name'  => $message->getName(),
+                'msg'   => $e->getMessage()
+            ]);
+            return false;
         }
     }
 }