+++ src/Consumer.php
@@ -24,17 +24,22 @@ interface Consumer
      * Run the consumer for a given queue. This will block.
      *
      * @param   string $queueName The queue from which the jobs will be consumed.
-     * @return  void
+     * @return  int The exit code to be used for the consumer.
      */
     public function run($queueName);
 
     /**
      * Consume a single job from the given queue. This will block until the
-     * job is competed then return.
+     * job is competed then return. Implementations of this method MUST be
+     * safe to run in a loop.
      *
      * @param   string $queueName The queue from which jobs will be consumed.
-     * @throws  Exception if anything goes wrong
-     * @return  void
+     * @throws  Exception\MustStop if the executor or handler throws a must
+     *          stop execption indicating a graceful stop is necessary
+     * @throws  Exception\DriverError|Exception if anything goes wrong with the
+     *          underlying driver itself.
+     * @return  boolean|null True if the a job was execute successfully. Null if
+     *          no job was executed. See the logs.
      */
     public function once($queueName);