+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+All notable changes to this project will be documented in this file.
+This project adheres to [Semantic Versioning](http://semver.org/).
+
+## 3.0.0 (Unreleased)
+
+### Changed
+
+- [BC BREAK] PHP version requirement was bumped to 5.6
+- [BC BREAK] `DefaultConsumer::once` (and the `Consumer` interface) have been
+  changed to make `once` safe to run in a loop. In other words, it never throws
+  exceptions unless it's a must stop or an exception thrown from a driver. All
+  other exceptions are logged, but not fatal. The idea here is to make consumers
+  safe to decorate without having to duplicate the error handling logic.
+- `Consumer` has docblocks that reflect its actual return values now.
+
+### Fixed
+
+n/a
+
+### Added
+
+- An `AbstractConsumer` class that provides the `run` and `stop` methods for
+  consumers without tying them to a specific implementation of `once`.
+
 ## 2.0.2
 
 Simple license update.